Remove MSVC scripts
This commit removes all the scripts located in src/tools/msvc/ to build
PostgreSQL with Visual Studio on Windows, meson becoming the recommended
way to achieve that.  The scripts held some information that is still
relevant with meson, information kept and moved to better locations.
Comments that referred directly to the scripts are removed.

All the documentation still relevant that was in install-windows.sgml
has been moved to installation.sgml under a new subsection for Visual.
All the content specific to the scripts is removed.  Some adjustments
for the documentation are planned in a follow-up set of changes.

Author: Michael Paquier
Reviewed-by: Peter Eisentraut, Andres Freund
Discussion: https://postgr.es/m/ZQzp_VMJcerM1Cs_@paquier.xyz

Conflicts:
	src/tools/msvc/Install.pm
	src/tools/msvc/MSBuildProject.pm
	src/tools/msvc/Mkvcbuild.pm
	src/tools/msvc/Project.pm
	src/tools/msvc/Solution.pm
	src/tools/msvc/VSObjectFactory.pm
	src/tools/msvc/build.pl
	src/tools/msvc/clean.bat
	src/tools/msvc/config_default.pl
	src/tools/msvc/dummylib/Win32/Registry.pm
	src/tools/msvc/dummylib/Win32API/File.pm
	src/tools/msvc/pgbison.pl
	src/tools/msvc/vcregress.pl