Use DELETE instead of UPDATE to speed up vacuum test
d42f60ccf07d89c introduced a test which generated dead tuples for vacuum
with an UPDATE. The test only required enough dead TIDs for two rounds
of index vacuuming. This can be accomplished with a DELETE instead of an
UPDATE -- which generates about 50% less WAL and makes the test 20%
faster in many cases. The test takes several seconds (more on slow
buildfarm animals) because we need quite a few tuples to trigger two
rounds of index vacuuming; so it is worth a follow-on commit to speed it
up.

Suggested-by: Masahiko Sawada
Discussion: https://postgr.es/m/CAAKRu_bWmMjmqL%2BOZ2duEQ80u7cRvpsExLNZNjzk-pXX5skwMQ%40mail.gmail.com
Backpatch-through: 14, the first version containing this test.